moves Person to utilities and updates references
